+++ libstdc++-v3/testsuite/29_atomics/atomic_float/value_init.cc
@@ -1,4 +1,4 @@
-// Copyright (C) 2020 Free Software Foundation, Inc.
+// Copyright (C) 2020-2021 Free Software Foundation, Inc.
 //
 // This file is part of the GNU ISO C++ Library.  This library is free
 // software; you can redistribute it and/or modify it under the
@@ -22,13 +22,13 @@
 #include <atomic>
 #include <testsuite_hooks.h>
 
-constexpr std::atomic<double> a;
+constexpr std::atomic<float> a;
 
 void
 test01()
 {
   VERIFY(a.load() == 0);
-  static_assert(std::is_nothrow_default_constructible_v<std::atomic<double>>);
+  static_assert(std::is_nothrow_default_constructible_v<std::atomic<float>>);
 }
 
 int
